Split HIR signature resolution (#1301)
Move function and method signature collection, trait impl signature validation, and HIR symbol-name resolution from hir.rs into hir/signatures.rs while preserving compiler.hir behavior. Update the compiler source monolith reporter and ratchet ceilings for the reduced HIR facade line count.

Refs #1254.
